requirements.txt
#https://docs.github.com/en/codespaces/setting-up-your-project-for-codespaces/adding-a-dev-container-configuration/setting-up-your-python-project-for-codespaces
# sudo apt-get update 
# sudo apt-get install libgl1

pytest==8.2.2
pytest-cov==5.0.0
pylint==3.2.3
black==24.4.2
ipython==8.18.1
 
joblib==1.3.2

#########################################
#######           azure             #####
#########################################
azure-ai-textanalytics
azure-ai-translation-text

#########################################
####### hugging face and poythorch  #####
#########################################
huggingface-hub==0.23.4
safetensors==0.4.3
tokenizers==0.19.1
transformers==4.41.2
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0
opencv-contrib-python==4.10.0.84
fastai
kagglehub

numpy
#==1.23.5
pandas==1.3.5


matplotlib
seaborn
plotly

scikit-learn==1.3.2
scikit-optimize==0.9.0 
scipy==1.11.3
statsmodels 
 
nbformat>=4.2.0

